몬스터 프리팹 에디터 제작중

+++ Styx_Station/Assets/03. Scripts/ScriptableObject/Stage.cs
@@ -12,7 +12,7 @@ public class Stage : ScriptableObject
     public float waveTimer;
     public int monsterAttackIncrease;
     public int monsterHealthIncrease;
-    public int monsterAttackSpeedIncrease;
+    public float monsterAttackSpeedIncrease;
     public MonsterTypeBase monster1;
     public int monster1Count;
     public MonsterTypeBase monster2;
